gic: seperate rec independent code from rec/gic.rs
Moved device-specific GIC config from rec/gic.rs
to the parent directory's.
Now, rec/gic.rs only contains REC context-specific code.
