+++ rmm/src/rmi/features.rs
@@ -1,7 +1,7 @@
 use crate::event::RmiHandle;
+use crate::gic;
 use crate::listen;
 use crate::rec;
-use crate::rec::gic;
 use crate::rmi;
 use crate::simd;
 use armv9a::{define_bitfield, define_bits, define_mask};
